+++ custom_components/gazpar/enum.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+# ----------------------------
+class FrequencyStr(Enum):
+    HOURLY = "hourly"
+    DAILY = "daily"
+    WEEKLY = "weekly"
+    MONTHLY = "monthly"
+    YEARLY = "yearly"
+
+    def __str__(self):
+        return self.value